add js tests
